Fixed DNS Server
DNS server technically was working before. It could receive messages and
send responses. But it would unfortunately send incorrect/malformed
responses. This is because the packer and unpacker of the flags was
broken. I fixed that. Additionally, this caused issues where the server
would try to continuously parse udp. Not sure what was wrong there but
its fixed now. Added some tests for the packer and unpacker.